+++ Controls-demo/list_new/ItemActions/ContextMenuConfig/FooterTemplate/FooterTemplate.less
@@ -1,0 +1,6 @@
+@import "../../../../../Controls-default-theme/_theme";
+
+.ControlsDemo_MenuFooterTemplate-footer {
+  padding: 10px 15px;
+  background-color: @primary_color;
+}
