+++ Controls-demo/list_new/ItemActions/ContextMenuConfig/HeaderTemplate/HeaderTemplate.less
@@ -1,0 +1,6 @@
+@import "../../../../../Controls-default-theme/_theme";
+
+.ControlsDemo_MenuHeaderTemplate-header {
+  padding: 32px 15px 10px;
+  background-color: @primary_color;
+}
