+++ Controls-demo/list_new/ItemActions/ContextMenuConfig/HeaderTemplate/HeaderTemplate.wml
@@ -1,0 +1,14 @@
+<div class="controlsDemo__wrapper controlsDemo_fixedWidth500 controlsDemo__itemActions_list_HeaderTemplate">
+    <Controls.list:View
+            keyProperty="id"
+            source="{{_viewSource}}"
+            itemActions="{{_itemActions}}">
+        <ws:contextMenuConfig>
+            <ws:headerTemplate>
+                <div class="ControlsDemo_MenuHeaderTemplate-header">
+                    В шаблоне headerTemplate может быть<br/>размещена дополнительная информация
+                </div>
+            </ws:headerTemplate>
+        </ws:contextMenuConfig>
+    </Controls.list:View>
+</div>
