+++ Controls-demo/list_new/ItemActions/ContextMenuConfig/Index.ts
@@ -1,0 +1,8 @@
+import {Control, TemplateFunction} from 'UI/Base';
+import * as Template from 'wml!Controls-demo/list_new/ItemActions/ContextMenuConfig/ContextMenuConfig';
+
+export default class extends Control {
+    protected _template: TemplateFunction = Template;
+
+    static _styles: string[] = ['Controls-demo/Controls-demo'];
+}
