+++ Controls-demo/list_new/ItemActions/ContextMenuConfig/ItemTemplate/ItemTemplate.less
@@ -1,4 +1,4 @@
-@import "Controls-default-theme/_theme";
+@import "../../../../../Controls-default-theme/_theme";
 
 .ControlsDemo_MenuItemTemplate-download-menu-item__size {
   color: @label_text-color;
